Move `string_enum` to `util` module
